Added an integration test for the globs feature
